refactor: migrate `/utxo/inscriptions` API calls to v2
